order/permissions.py
from rest_framework import permissions


class IsOwner(permissions.BasePermission):
    def has_object_permission(self, request, view, obj):
        if request.user.is_staf:
            return True
        return obj.shipper.user == request.user
