requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.3
django-environ==0.10.0
django-filter==23.2
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
drf-spectacular==0.26.3
drf-spectacular-sidecar==2023.7.1
importlib-metadata==6.7.0
inflection==0.5.1
jsonschema==4.17.3
Markdown==3.4.3
phonenumbers==8.13.15
Pillow==10.0.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
typing-extensions==4.7.1
uritemplate==4.1.1
zipp==3.15.0
